move solving-related functions to solve.py
