+++ src/halmos/constants.py
@@ -1,0 +1,4 @@
+VERBOSITY_TRACE_COUNTEREXAMPLE = 2
+VERBOSITY_TRACE_SETUP = 3
+VERBOSITY_TRACE_PATHS = 4
+VERBOSITY_TRACE_CONSTRUCTOR = 5
